Go-Kart Automation Web Application
The Go-Kart Automation Web Application is a software system designed to enhance the functionality and safety of go-karts. It provides various features such as wireless starting and stopping of the kart, kill switch status indication, driver safety intimation from the team, and many more. This README provides an overview of the application and instructions for installation and usage.